services/mock_audio_bridge/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
paho-mqtt>=1.6.1
pydantic>=2.0.0
